description: render as markdown
Prior to this change the description parameter was templated as a
string. This made it difficult to insert links or make text bold, as the
Go template engine escapes any html markup.

After this change the string is rendered with .RenderString which uses
the default page markup, markdown, to render the string.